refactor(cli): flatten generate command module
Move the Generate command slice to the neutral command root and update every internal caller without changing public CLI or generation behavior.
