+++ turbo/apps/cli/src/commands/generate/__tests__/artifacts.test.ts
