+++ turbo/apps/cli/src/commands/generate/__tests__/sprite.test.ts
@@ -24,7 +24,7 @@ describe("okou generate sprite command", () => {
   beforeEach(() => {
     chalk.level = 0;
     vi.stubEnv("VM0_API_BACKEND_URL", "http://localhost:3000");
-    vi.stubEnv("OKOU_TOKEN", "test-zero-token");
+    vi.stubEnv("OKOU_TOKEN", "test-token");
   });
 
   afterEach(() => {
