+++ turbo/apps/cli/src/commands/generate/image.ts
@@ -1,4 +1,4 @@
-import { createImageGenerateCommand } from "../../shared/image-generate";
+import { createImageGenerateCommand } from "../shared/image-generate";
 
 export const imageCommand = createImageGenerateCommand({
   name: "image",
