+++ turbo/apps/cli/src/commands/generate/sprite.ts
@@ -1,9 +1,9 @@
 import { Command } from "commander";
-import { withErrorHandler } from "../../../lib/command/with-error-handler";
+import { withErrorHandler } from "../../lib/command/with-error-handler";
 import {
   createSpriteAuthoringPacket,
   type SpritePlan,
-} from "../../shared/sprite-authoring";
+} from "../shared/sprite-authoring";
 import { dispatchGenerate } from "./lib/dispatch";
 
 const SPRITE_USAGE_COMMAND = "okou generate sprite";
