+++ turbo/apps/cli/src/commands/shared/image-generate.ts
@@ -8,8 +8,8 @@ import {
   listImageStyles,
 } from "@okouai/core/resource-registry";
 import { formatRegistryListing } from "./resource-listing";
-import { dispatchGenerate } from "../zero/generate/lib/dispatch";
-import type { GenerationType } from "../zero/generate/lib/lister";
+import { dispatchGenerate } from "../generate/lib/dispatch";
+import type { GenerationType } from "../generate/lib/lister";
 
 interface ImageOptions {
   prompt?: string;
