+++ turbo/apps/cli/src/commands/shared/presentation-generate.ts
@@ -1,7 +1,7 @@
 import { Command, InvalidArgumentError } from "commander";
 import { withErrorHandler } from "../../lib/command/with-error-handler";
-import { dispatchGenerate } from "../zero/generate/lib/dispatch";
-import type { GenerationType } from "../zero/generate/lib/lister";
+import { dispatchGenerate } from "../generate/lib/dispatch";
+import type { GenerationType } from "../generate/lib/lister";
 import {
   buildPresentationRunbookInstructionLines,
   findPresentationRunbookPackage,
