+++ turbo/apps/cli/src/commands/shared/video-generate.ts
@@ -14,8 +14,8 @@ import {
   currentPlanAllowsVideo,
   currentTokenCanReadBilling,
 } from "./billing-capabilities";
-import { dispatchGenerate } from "../zero/generate/lib/dispatch";
-import type { GenerationType } from "../zero/generate/lib/lister";
+import { dispatchGenerate } from "../generate/lib/dispatch";
+import type { GenerationType } from "../generate/lib/lister";
 
 interface VideoOptions {
   prompt?: string;
